Move TestHiveParquetWithBloomFilters to io.trino.testing
